+++ app/src/main/java/com/nucldev/simpleweatherlocator/netinteraction/forecastpojo/Main.java
@@ -13,19 +13,19 @@ public class Main{
 	private double tempMin;
 
 	@SerializedName("grnd_level")
-	private int grndLevel;
+	private double grndLevel;
 
 	@SerializedName("temp_kf")
-	private int tempKf;
+	private double tempKf;
 
 	@SerializedName("humidity")
-	private int humidity;
+	private double humidity;
 
 	@SerializedName("pressure")
-	private int pressure;
+	private double pressure;
 
 	@SerializedName("sea_level")
-	private int seaLevel;
+	private double seaLevel;
 
 	@SerializedName("feels_like")
 	private double feelsLike;
@@ -53,7 +53,7 @@ public class Main{
 		this.grndLevel = grndLevel;
 	}
 
-	public int getGrndLevel(){
+	public double getGrndLevel(){
 		return grndLevel;
 	}
 
@@ -61,7 +61,7 @@ public class Main{
 		this.tempKf = tempKf;
 	}
 
-	public int getTempKf(){
+	public double getTempKf(){
 		return tempKf;
 	}
 
@@ -69,7 +69,7 @@ public class Main{
 		this.humidity = humidity;
 	}
 
-	public int getHumidity(){
+	public double getHumidity(){
 		return humidity;
 	}
 
@@ -77,7 +77,7 @@ public class Main{
 		this.pressure = pressure;
 	}
 
-	public int getPressure(){
+	public double getPressure(){
 		return pressure;
 	}
 
@@ -85,7 +85,7 @@ public class Main{
 		this.seaLevel = seaLevel;
 	}
 
-	public int getSeaLevel(){
+	public double getSeaLevel(){
 		return seaLevel;
 	}
